Taggregator: Add support for Spotify
This commit introduces three more properties a site object can have.
The first, `needsAuthentication`, should be a function that returns
true when the site needs some form of authentication in order to grab
its tags. The second, `authenticate`, is a function that opens a new
window or tab, initiates authorization, and redirects back to the
musicbrainz homepage of choice (either musicbrainz.org,
musicbrainz.eu, beta.musicbrainz.org, or test.musicbrainz.org). The
third, `redirectHandler`, is called after the authorization is
complete on the homepage when the `taggregator-auth` URL search
parameter matches the site object's domain and is responsible for
storing the authentication. All three functions are called with no
arguments.