Taggregator: Allow importing tags from individial sites
Adds 2 required fields to the site object: `name`, is a
human-readable name for the site to be displayed in the settings.
`favicon-class` is the class the MBz website uses to show the favicon
of the site in the external_links list.

This commit does introduce a conflict with Lioncat6's SAMBL-Button
userscript. Because we're injecting an element with the
`spotify-favicon` class, the SAMBL userscript looks for a link that
isn't there and errors.

Fixes #28.